refactor: reorganize provider-specific funcions
Move provider-specific methods out of http_utils into their relevant
client file.
