+++ src/linux_mcp_server/gatekeeper/openrouter_client.py
@@ -1,16 +1,48 @@
 """OpenRouter Chat Completions client for the gatekeeper."""
 
+import os
+
 from typing import Any
 
 from linux_mcp_server.config import CONFIG
+from linux_mcp_server.config import ReasoningEffort
 from linux_mcp_server.gatekeeper.http_utils import DEFAULT_TIMEOUT_SECONDS
-from linux_mcp_server.gatekeeper.http_utils import get_openrouter_base_url
-from linux_mcp_server.gatekeeper.http_utils import normalize_openrouter_model_id
-from linux_mcp_server.gatekeeper.http_utils import openrouter_auth_headers
-from linux_mcp_server.gatekeeper.http_utils import openrouter_reasoning_block
 from linux_mcp_server.gatekeeper.http_utils import post_json
-from linux_mcp_server.gatekeeper.llm import GatekeeperCompletion
 from linux_mcp_server.gatekeeper.schema import openai_response_format
+from linux_mcp_server.models import GatekeeperCompletion
+
+
+OPENROUTER_DEFAULT_BASE_URL = "https://openrouter.ai/api/v1"
+
+
+def _normalize_openrouter_model_id(model: str) -> str:
+    if model.startswith("openrouter/"):
+        return model[len("openrouter/") :]
+    return model
+
+
+def _get_openrouter_base_url() -> str:
+    configured = CONFIG.gatekeeper.openrouter.base_url if CONFIG.gatekeeper.openrouter else None
+    return (configured or OPENROUTER_DEFAULT_BASE_URL).rstrip("/")
+
+
+def _get_openrouter_api_key() -> str:
+    api_key = os.environ.get("OPENROUTER_API_KEY")
+    if not api_key:
+        raise ValueError("OPENROUTER_API_KEY is required for OpenRouter gatekeeper provider.")
+    return api_key
+
+
+def _openrouter_auth_headers() -> dict[str, str]:
+    return {"Authorization": f"Bearer {_get_openrouter_api_key()}"}
+
+
+def _openrouter_reasoning_block(reasoning_effort: ReasoningEffort | None) -> dict[str, Any] | None:
+    if reasoning_effort is None or reasoning_effort == ReasoningEffort.DEFAULT:
+        return None
+    if reasoning_effort == ReasoningEffort.NONE:
+        return {"enabled": False}
+    return {"enabled": True, "effort": reasoning_effort.value}
 
 
 def _openrouter_config() -> dict[str, Any]:
@@ -29,14 +61,14 @@ def _build_chat_completions_body(prompt: str) -> dict[str, Any]:
         provider["quantizations"] = [openrouter["quantization"]]
 
     body: dict[str, Any] = {
-        "model": normalize_openrouter_model_id(CONFIG.gatekeeper.model or ""),
+        "model": _normalize_openrouter_model_id(CONFIG.gatekeeper.model or ""),
         "messages": [{"role": "user", "content": prompt}],
         "temperature": CONFIG.gatekeeper.temperature,
         "provider": provider,
     }
     if CONFIG.gatekeeper.structured_output:
         body["response_format"] = openai_response_format()
-    reasoning = openrouter_reasoning_block(CONFIG.gatekeeper.reasoning_effort)
+    reasoning = _openrouter_reasoning_block(CONFIG.gatekeeper.reasoning_effort)
     if reasoning is not None:
         body["reasoning"] = reasoning
     if openrouter["template_kwargs"]:
@@ -68,9 +100,9 @@ def _extract_usage(response: dict[str, Any]) -> tuple[int, int, float | None]:
 
 
 def complete_openrouter(prompt: str, *, timeout: int = DEFAULT_TIMEOUT_SECONDS) -> GatekeeperCompletion:
-    base_url = get_openrouter_base_url()
+    base_url = _get_openrouter_base_url()
     headers = {
-        **openrouter_auth_headers(),
+        **_openrouter_auth_headers(),
         "Content-Type": "application/json",
     }
     response = post_json(
