+++ src/linux_mcp_server/gatekeeper/vertex_ai_client.py
@@ -1,5 +1,7 @@
 """Vertex AI gatekeeper client with model-based API routing."""
 
+from typing import Literal
+
 from linux_mcp_server.config import CONFIG
 from linux_mcp_server.gatekeeper.anthropic_client import build_messages_body
 from linux_mcp_server.gatekeeper.anthropic_client import extract_messages_text
@@ -7,16 +9,42 @@ from linux_mcp_server.gatekeeper.gcp_auth import get_gcp_location
 from linux_mcp_server.gatekeeper.gcp_auth import get_gcp_project
 from linux_mcp_server.gatekeeper.gemini_client import build_gemini_body
 from linux_mcp_server.gatekeeper.gemini_client import extract_gemini_text
-from linux_mcp_server.gatekeeper.http_utils import ANTHROPIC_VERTEX_VERSION
 from linux_mcp_server.gatekeeper.http_utils import DEFAULT_TIMEOUT_SECONDS
-from linux_mcp_server.gatekeeper.http_utils import get_vertex_openapi_base_url
 from linux_mcp_server.gatekeeper.http_utils import normalize_model_id
 from linux_mcp_server.gatekeeper.http_utils import post_json
-from linux_mcp_server.gatekeeper.http_utils import vertex_api_style
-from linux_mcp_server.gatekeeper.http_utils import vertex_auth_headers
-from linux_mcp_server.gatekeeper.llm import GatekeeperCompletion
 from linux_mcp_server.gatekeeper.openai_client import build_chat_completions_body
 from linux_mcp_server.gatekeeper.openai_client import extract_chat_completions_text
+from linux_mcp_server.models import GatekeeperCompletion
+
+
+ANTHROPIC_VERTEX_VERSION = "vertex-2023-10-16"
+
+
+def _vertex_api_style(model: str) -> Literal["anthropic", "gemini", "openai_compatible"]:
+    normalized = normalize_model_id(model)
+    if normalized.startswith("claude"):
+        return "anthropic"
+    if normalized.startswith("gemini"):
+        return "gemini"
+    return "openai_compatible"
+
+
+def _get_vertex_openapi_base_url() -> str:
+    cfg = CONFIG.gatekeeper.vertex_ai
+    if cfg and cfg.base_url:
+        return cfg.base_url.rstrip("/")
+    from linux_mcp_server.gatekeeper.gcp_auth import get_gcp_location
+    from linux_mcp_server.gatekeeper.gcp_auth import get_gcp_project
+
+    project = get_gcp_project()
+    location = get_gcp_location()
+    return f"https://aiplatform.googleapis.com/v1/projects/{project}/locations/{location}/endpoints/openapi"
+
+
+def _vertex_auth_headers() -> dict[str, str]:
+    from linux_mcp_server.gatekeeper.gcp_auth import get_gcp_access_token
+
+    return {"Authorization": f"Bearer {get_gcp_access_token()}"}
 
 
 def _anthropic_vertex_url(model: str) -> str:
@@ -40,7 +68,7 @@ def _complete_anthropic_on_vertex(prompt: str, *, timeout: int) -> GatekeeperCom
     response = post_json(
         provider="anthropic",
         url=_anthropic_vertex_url(model),
-        headers={**vertex_auth_headers(), "Content-Type": "application/json"},
+        headers={**_vertex_auth_headers(), "Content-Type": "application/json"},
         body=body,
         timeout=timeout,
     )
@@ -52,7 +80,7 @@ def _complete_gemini_on_vertex(prompt: str, *, timeout: int) -> GatekeeperComple
     response = post_json(
         provider="gemini",
         url=_gemini_vertex_url(model),
-        headers={**vertex_auth_headers(), "Content-Type": "application/json"},
+        headers={**_vertex_auth_headers(), "Content-Type": "application/json"},
         body=build_gemini_body(prompt),
         timeout=timeout,
     )
@@ -60,11 +88,11 @@ def _complete_gemini_on_vertex(prompt: str, *, timeout: int) -> GatekeeperComple
 
 
 def _complete_openai_compatible_on_vertex(prompt: str, *, timeout: int) -> GatekeeperCompletion:
-    base_url = get_vertex_openapi_base_url()
+    base_url = _get_vertex_openapi_base_url()
     response = post_json(
         provider="openai",
         url=f"{base_url}/chat/completions",
-        headers={**vertex_auth_headers(), "Content-Type": "application/json"},
+        headers={**_vertex_auth_headers(), "Content-Type": "application/json"},
         body=build_chat_completions_body(prompt),
         timeout=timeout,
     )
@@ -73,12 +101,10 @@ def _complete_openai_compatible_on_vertex(prompt: str, *, timeout: int) -> Gatek
 
 def complete_vertex_ai(prompt: str, *, timeout: int = DEFAULT_TIMEOUT_SECONDS) -> GatekeeperCompletion:
     model = CONFIG.gatekeeper.model or ""
-    match vertex_api_style(model):
+    match _vertex_api_style(model):
         case "anthropic":
             return _complete_anthropic_on_vertex(prompt, timeout=timeout)
         case "gemini":
             return _complete_gemini_on_vertex(prompt, timeout=timeout)
         case "openai_compatible":
             return _complete_openai_compatible_on_vertex(prompt, timeout=timeout)
-        case _:  # pragma: no cover
-            raise ValueError(f"Unsupported Vertex AI API style for model: {model}")
