+++ src/linux_mcp_server/models.py
@@ -198,3 +198,10 @@ class LogEntries(BaseModel):
     @field_serializer("unit", "path")
     def serialize_empty_as_null(self, value: str | Path | None) -> str | None:
         return str(value) if value else None
+
+
+class GatekeeperCompletion(BaseModel):
+    text: str
+    prompt_tokens: int = 0
+    completion_tokens: int = 0
+    usage_cost: float | None = None
